--- Assets/Scripts/RifleNew.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3fe763604ca655c4caad1307067468f2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
