+++ Assets/Scripts/UI/AimUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a6548115808062408a5b7b440a7a350
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
